Update state machine and add unit test.

+++ hpm-esp/hpma115s.h
@@ -14,19 +14,18 @@ typedef enum {
 
 class HPMA115S {
 
-private:
-    EState      _state;
-    uint8_t     _sum;
-    uint8_t     _buf[32];
-    uint16_t    _idx;
-    uint16_t    _len;
-    uint16_t    _chk;
+  private:
+    EState _state;
+    int _sum;
+    uint8_t _buf[32];
+    size_t _idx;
+    int _len;
+    int _chk;
 
-public:
-    HPMA115S(void);
+  public:
+     HPMA115S(void);
     void reset(void);
     bool process(uint8_t c);
 
 };
-
